+++ R/describe.R
@@ -30,10 +30,10 @@ rws_describe_meta <- function(x, ..., conn) {
 #' rws_read_meta(conn)
 #' rws_disconnect(conn)
 rws_describe_meta.character <- function(x, column, description, ..., conn) {
-  chk_is(x, "character")
+  chk_is(x, "character"); chk_no_missing(x)
   chk_is(column, "character"); chk_no_missing(column)
   chk_is(description, "character")
-  check_sqlite_connection(conn, connected = TRUE)
+  chk_sqlite_conn(conn, connected = TRUE)
   chk_unused(...)
   
   rws_describe_meta(data.frame(
@@ -49,10 +49,15 @@ rws_describe_meta.character <- function(x, column, description, ..., conn) {
 #' @family rws_read
 #' @export
 rws_describe_meta.data.frame <- function(x, ..., conn) {
-  check_data(x, values = list(Table = "", Column = "", Description = c("", NA)))
-  check_sqlite_connection(conn, connected = TRUE)
-  chk_unused(...)
-  
+  if(is_chk_on()) {
+    chk_is(x, "data.frame")
+    chk_has(colnames(x), c("Table", "Column", "Description"));
+    chk_is(x$Table, "character"); chk_no_missing(x$Table)
+    chk_is(x$Column, "character"); chk_no_missing(x$Column)
+    chk_is(x$Description, "character")
+    chk_sqlite_conn(conn, connected = TRUE)
+    chk_unused(...)
+  }
   if(!nrow(x)) return(invisible(rws_read_meta(conn)))
   
   x <- x[c("Table", "Column", "Description")]
@@ -66,7 +71,7 @@ rws_describe_meta.data.frame <- function(x, ..., conn) {
   
   meta <- rws_read_meta(conn)
   meta$RowMeta <- 1:nrow(meta)
-
+  
   meta <- merge(meta, x, by.x = c("TableMeta", "ColumnMeta"), 
                 by.y = c("Table", "Column"), all = TRUE)