+++ R/namespace.R
@@ -1,4 +1,4 @@
-#' @import checkr chk RSQLite
+#' @import chk RSQLite
 #' @importFrom hms is_hms as_hms
 #' @importFrom dttr2 is.Date is.POSIXct
 NULL