+++ R/validate.R
@@ -7,7 +7,7 @@ validate_data <- function(data, table_name, strict, silent, conn) {
   names(data_names) <- to_upper(names(data))
   names(data) <- to_upper(names(data))
   
-  check_colnames(data, colnames = colnames, x_name = p0("data '", table_name, "'"))
+  chk_has(colnames(data), colnames, x_name = p0("'", table_name, "' column names"))
   
   if(isFALSE(silent)) {
     extra <- data_names[!names(data_names) %in% colnames]
@@ -33,7 +33,7 @@ validate_data <- function(data, table_name, strict, silent, conn) {
   }
   
   pk <- table_info$name[table_info$pk != 0L]
-  check_key(data, key = pk, x_name = p0("data '", table_name, "'"), 
+  checkr::check_key(data, key = pk, x_name = p0("data '", table_name, "'"), 
             na_distinct = TRUE)
   names(data) <- data_names[names(data)]
   if(!is.na(sf_column_name) && sf_column_name %in% names(data)) {