+++ tests/testthat/test-connection.R
@@ -4,7 +4,7 @@ test_that("rws_connect", {
   expect_error(rws_connect(":memory:", exists = TRUE),
                "File ':memory:' must already exist.")
   conn <- rws_connect(":memory:")
-  expect_identical(check_sqlite_connection(conn, connected = TRUE), conn)
+  expect_true(chk_sqlite_conn(conn, connected = TRUE))
   rws_disconnect(conn)
-  expect_identical(check_sqlite_connection(conn, connected = FALSE), conn)
+  expect_true(chk_sqlite_conn(conn, connected = FALSE))
 })